Test improvements
Add unit tests, refactor mocks, mock a resampled multiscene.
